Extract `DictionaryAdapter`
